I have quickstart working now. I had some consistent execution issues with it per other student review. There were some datatype requirements. I still need to update the quickstart doc with the updated script, but I have the updated script together.

+++ free_range_zoo/envs/rideshare/baselines/noop.py
@@ -13,7 +13,7 @@ class NoopBaseline(Agent):
         """Initialize the agent."""
         super().__init__(*args, **kwargs)
 
-        self.actions = torch.zeros((self.parallel_envs, 2), dtype=torch.int32)
+        self.actions = torch.zeros((self.parallel_envs, 2), dtype=torch.int16)
 
     def act(self, action_space: free_range_rust.Space) -> List[List[int]]:
         """
